requirements.txt
langchain
langchain-openai
streamlit
python-dotenv
langchain-anthropic 
python-pptx
python-docx
fastapi
uvicorn
